+++ src/packages/piral-core/src/components/components.tsx
@@ -5,7 +5,7 @@ import { ComponentsState, ErrorInfoProps, LoadingIndicatorProps, RouterProps, La
 export function getPiralComponent<TKey extends keyof ComponentsState>(name: TKey): ComponentsState[TKey] {
   return props => {
     const Component = useGlobalState(s => s.components[name]);
-    return <Component {...props} />;
+    return Component ? <Component {...props} /> : null;
   };
 }
 
@@ -13,3 +13,4 @@ export const PiralError: React.ComponentType<ErrorInfoProps> = getPiralComponent
 export const PiralLoadingIndicator: React.ComponentType<LoadingIndicatorProps> = getPiralComponent('LoadingIndicator');
 export const PiralRouter: React.ComponentType<RouterProps> = getPiralComponent('Router');
 export const PiralLayout: React.ComponentType<LayoutProps> = getPiralComponent('Layout');
+export const PiralDebug: React.ComponentType = getPiralComponent('Debug');