+++ src/packages/piral-core/src/helpers.tsx
@@ -71,7 +71,8 @@ export function createPiletOptions({
       },
       pilets: {
         createApi,
-        getDependencies,
+        getDependencies: getDependencyResolver(globalDependencies, getDependencies),
+        loadPilet,
         requestPilets,
       },
     };